requirements.txt
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
gptservice==0.1.0